% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R.R
\docType{data}
\name{dat_organization}
\alias{dat_organization}
\title{Data on organizational survey (Eid & Rauber, 2000)}
\format{
a data frame with 20 rows and 2 columns
\describe{
\item{sex}{sex}
\item{age}{age}
\item{length_service}{length of service}
\item{same_position}{length of service on the same position}
\item{item_1 to item_6}{item responses}
\item{class}{latent class (based on the original analysis by Eid & Rauber)}
\item{leadership}{leadership level}
}
}
\source{
Eid, M. & Rauber, M. (200). Detecting measurement invariance in organizational surveys.
European Journal of Psychological Assessment, 16, 20-30.
data from: http://www.hhpub.com/journals/ejpa
}
\usage{
dat_organization
}
\description{
Data used in the article by Eid & Rauber
}
\keyword{datasets}
