% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effecttree_PACKAGE.R
\name{effecttree-package}
\alias{effecttree-package}
\alias{effecttree}
\title{effecttree: Helper functions to use effect size measures
as a stopping criterion in model-based recursive partitioning trees}
\description{
Helper functions to use the Mantel-Haenszel effect size measure
and the partial gamma coefficient as a stopping criterion in Rasch trees and
partial credit trees
}
\author{
Mirka Henninger & Jan Radek
}
