% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_pgamma.R
\name{get_ABC_classification}
\alias{get_ABC_classification}
\title{Performs classification of the partial gamma effect size based on Bjorner et al. (1998)}
\usage{
get_ABC_classification(res_pgamma, res_pgamma_se, p.adj, threshold)
}
\arguments{
\item{res_pgamma}{A vector or dataframe with the partial gamma coefficient for each item}

\item{res_pgamma_se}{A vector or dataframe with the standard error of partial gamma coefficient for each item}

\item{p.adj}{A character indicating the correction method for multiple testing. Options are "none", "bonferroni" and "fdr"}

\item{threshold}{The threshold of partial gamma above which items should be labeled as DIF/DSF items, default is .21 and .31}
}
\value{
A vector or dataframe with the corresponding A/B/C classification for each item
}
\description{
Performs classification of the partial gamma effect size based on Bjorner et al. (1998)
}
