% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_effectsize.R
\name{get_effectsize}
\alias{get_effectsize}
\title{Function that returns the effect size measure of a tree object}
\usage{
get_effectsize(object, purification, by = "node")
}
\arguments{
\item{object}{An object of type modelparty}

\item{purification}{A character indicating the type of purification ("none", "iterative")}

\item{by}{A character either "node" or "type" defining the structure of the function output}
}
\value{
If by = "node": a list with one entry for each node; If by = "type": a list with entries effect size, classification, purification type, and purification_counter
}
\description{
Function that returns the effect size measure of a tree object
}
\examples{
\dontrun{
data("DIFSim", package = "psychotree")
RT <- raschtree(resp ~ age + gender + motivation, data = DIFSim)
eff <- get_effectsize(RT, purification = "iterative", by = "type")
eff
}

}
