% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prune_nodes.R
\name{get_prune_nodes}
\alias{get_prune_nodes}
\title{Extract the inner nodes to be pruned from the effect size measure}
\usage{
get_prune_nodes(object, direction, evalcrit)
}
\arguments{
\item{object}{An object of type effecttree}

\item{direction}{A character either "topdown" or "bottomup" indicating whether stopping (topdown) or pruning (bottomup) should be performed}

\item{evalcrit}{A character or character vector indicating the evalution criterion. The default is "A", for which all nodes are pruned in which all items are categorized as "A"}
}
\value{
A vector with the number of the nodes to be pruned
}
\description{
Extract the inner nodes to be pruned from the effect size measure
}
