% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_pgamma.R
\name{purify_pgamma}
\alias{purify_pgamma}
\title{Performs purification of the partial gamma coefficient}
\usage{
purify_pgamma(dat, group, res_gamma, threshold)
}
\arguments{
\item{dat}{A dataframe with polytomous item responses}

\item{group}{An indicator of the two groups for which DIF/DSF analyses should be performed}

\item{res_gamma}{A vector with partial gamma coefficient}

\item{threshold}{The threshold of partial gamma above which items should be labeled as DIF items}
}
\value{
A dataframe including purified partial gamma coefficients
}
\description{
Performs purification of the partial gamma coefficient
}
