% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_effectsize.R
\name{show_effectsize}
\alias{show_effectsize}
\title{Helper function to display the effect size measure and classification summary
in each inner node of the tree}
\usage{
show_effectsize(
  object,
  id = TRUE,
  pval = TRUE,
  abbreviate = FALSE,
  fill = "white",
  gp = gpar()
)
}
\arguments{
\item{object}{An object of class modelparty that has the effect size measure added to it. See @examples}

\item{id}{Argument copied from partykit:::node_inner}

\item{pval}{Argument copied from partykit:::node_inner}

\item{abbreviate}{Argument copied from partykit:::node_inner}

\item{fill}{Argument copied from partykit:::node_inner}

\item{gp}{Argument copied from partykit:::node_inner}
}
\value{
A function that can be used as a value for the argument 'inner_panel' in plot.raschtree() or plot.pctree()
}
\description{
This function is largely copied from partykit:::node_inner and
adapted to display larger inner nodes with corresponding effect size classification summaries
}
