% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_effectsize.R
\name{add_effectsize}
\alias{add_effectsize}
\title{Creates a effecttree object based on the original tree with additional effect size information}
\usage{
add_effectsize(object, type, purification)
}
\arguments{
\item{object}{An object of type modelparty}

\item{type}{A character indicating the type of the tree ("raschtree", "pctree")}

\item{purification}{A character indicating the type of purification ("none", "iterative")
############# FIXME ############ add correction for multiple testing}
}
\value{
An object of type modelparty with 'info' extended by a list named effectsize containing entries effect size, classification, purification type, and purificationCounter
}
\description{
Creates a effecttree object based on the original tree with additional effect size information
}
\examples{
\dontrun{
data("DIFSim", package = "psychotree")
RT <- raschtree(resp ~ age + gender + motivation, data = DIFSim)
RT_MH <- add_effectsize(RT, type = "raschtree", purification = "iterative")
RT_MH$info$effectsize
plot(RT_MH, color_by_node = 1)
}
}
