% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_effectsize.R
\name{get_effectsize}
\alias{get_effectsize}
\title{Function that returns the effect size measure of a tree object}
\usage{
get_effectsize(object, model, purification, p.adj, threshold = c(0.21, 0.31))
}
\arguments{
\item{object}{An object of type modelparty}

\item{model}{A character indicating the model in the tree object ("raschtree", "pctree")}

\item{purification}{A character indicating the type of purification ("none", "2step", "iterative")}

\item{p.adj}{A character indicating the correction method for multiple testing. Options are "none", "bonferroni" and "fdr"}

\item{threshold}{The threshold of partial gamma above which items should be labeled as DIF/DSF items, default is .21 and .31}
}
\value{
A list with entries effect size, classification, purification type, and purification_counter for each inner node
}
\description{
Function that returns the effect size measure of a tree object
}
\examples{
\dontrun{
## raschtree
data("DIFSim", package = "psychotree")
RT <- raschtree(resp ~ age + gender + motivation, data = DIFSim)
eff <- get_effectsize(RT, model = "raschtree", purification = "iterative")
eff

## RStree
data("dat_organization")
items <- as.data.frame(dat_organization[, grep("item", names(dat_organization))])
dat_organization$items <- apply(items,2,as.numeric)
rs_tree <- rstree(items ~ sex + age + length_service + same_position + leadership,
                  data = dat_organization)
eff <- get_effectsize(rs_tree, model = "rstree", p.adj = "fdr", purification = "iterative")
eff
}

}
