% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_effectsize.R
\name{get_effectsize}
\alias{get_effectsize}
\title{Function that returns the effect size measure of a tree object}
\usage{
get_effectsize(object, type, purification)
}
\arguments{
\item{object}{An object of type modelparty}

\item{type}{A character indicating the type of the tree object ("raschtree", "pctree")}

\item{purification}{A character indicating the type of purification ("none", "iterative")}
}
\value{
A list with entries effect size, classification, purification type, and purification_counter for each inner node
}
\description{
Function that returns the effect size measure of a tree object
}
\examples{
\dontrun{
data("DIFSim", package = "psychotree")
RT <- raschtree(resp ~ age + gender + motivation, data = DIFSim)
eff <- get_effectsize(RT, type = "raschtree", purification = "iterative")
eff
}

}
